Move screen clearing to system for testing purposes.
